+++ packages/terra-spacer/src/terra-dev-site/doc/spacer/Spacer.doc.jsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import DocTemplate from 'terra-doc-template';
-import ReadMe from '../../../docs/README.md';
-import { name } from '../../../package.json';
+import ReadMe from '../../../../docs/README.md';
+import { name } from '../../../../package.json';
 
 /* eslint-disable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 // Component Source
-import SignatureSrc from '!raw-loader!../../../src/Signature';
+import SpacerSrc from '!raw-loader!../../../../src/Spacer';
 
 // Example Files
-import SignatureExample from './example/SignatureExample';
-import SignatureExampleSrc from '!raw-loader!../../../src/terra-dev-site/doc/example/SignatureExample';
+import Spacer from '../example/SpacerExample';
+import SpacerExampleSrc from '!raw-loader!../../../../src/terra-dev-site/doc/example/SpacerExample';
 /* eslint-enable import/no-webpack-loader-syntax, import/first, import/extensions, import/no-unresolved, import/no-duplicates */
 
 const DocPage = () => (
@@ -19,15 +19,16 @@ const DocPage = () => (
     srcPath={`https://github.com/cerner/terra-core/tree/master/packages/${name}`}
     examples={[
       {
-        title: 'Default Signature',
-        example: <SignatureExample />,
-        source: SignatureExampleSrc,
+        title: 'Spacer',
+        description: 'Spacing default button with a padding value of large+4 and primary button with padding value of large+2',
+        example: <Spacer />,
+        source: SpacerExampleSrc,
       },
     ]}
     propsTables={[
       {
-        componentName: 'Signature',
-        componentSrc: SignatureSrc,
+        componentName: 'Spacer',
+        componentSrc: SpacerSrc,
       },
     ]}
   />
